Produce an NIH-compliant enrollment table.

Produce an NIH enrollment table, leveraging metadata to adapt to the observed data.frame.

table_nih_enrollment(
  d,
  d_lu_gender = NULL,
  d_lu_race = NULL,
  d_lu_ethnicity = NULL,
  variable_gender = "gender",
  variable_race = "race",
  variable_ethnicity = "ethnicity"
)

Arguments

d

data.frame of observed values in the investigation. Required.

d_lu_gender

data.frame that maps the observed levels of gender to the NIH-recommended levels of gender. Required only if the levels are not the same.

d_lu_race

data.frame that maps the observed levels of gender to the NIH-recommended levels of gender. Required only if the levels are not the same.

d_lu_ethnicity

data.frame that maps the observed levels of gender to the NIH-recommended levels of gender. Required only if the levels are not the same.

variable_gender

name of the gender variable in the d data.frame. Defaults to gender.

variable_race

name of the race variable in the d data.frame. Defaults to race.

variable_ethnicity

name of the ethnicity variable in the d data.frame. Defaults to ethnicity.

Value

Table for publication

Details

https://grants.nih.gov/grants/how-to-apply-application-guide/forms-d/general/g.500-phs-inclusion-enrollment-report.htm

Author

Will Beasley, Peter Higgins, Andrew Peters, Sreeharsha Mandem

Examples

library(magrittr)
ds_1 <- tibble::tribble(
  ~subject_id,   ~gender  , ~race                      ,   ~ethnicity                     ,
           1L,   "Male"   , "Black or African American",  "Not Hispanic or Latino"        ,
           2L,   "Male"   , "Black or African American",  "Not Hispanic or Latino"        ,
           3L,   "Female" , "Black or African American",  "Unknown/Not Reported Ethnicity",
           4L,   "Male"   , "White"                    ,  "Not Hispanic or Latino"        ,
           5L,   "Male"   , "White"                    ,  "Not Hispanic or Latino"        ,
           6L,   "Female" , "White"                    ,  "Not Hispanic or Latino"        ,
           7L,   "Male"   , "White"                    ,  "Hispanic or Latino"            ,
           8L,   "Male"   , "White"                    ,  "Hispanic or Latino"
)

table_nih_enrollment(ds_1)
#> Warning: `select_()` was deprecated in dplyr 0.7.0.
#> Please use `select()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_lifecycle_warnings()` to see where this warning was generated.
table_nih_enrollment_pretty(ds_1)
#> <table class="table table-striped table-hover table-condensed table-responsive" style="width: auto !important; margin-left: auto; margin-right: auto;">
#>  <thead>
#> <tr>
#> <th style="empty-cells: hide;border-bottom:hidden;" colspan="1"></th>
#> <th style="border-bottom:hidden;padding-bottom:0; padding-left:3px;padding-right:3px;text-align: center; " colspan="9"><div style="border-bottom: 1px solid #ddd; padding-bottom: 5px; ">Ethnic Categories</div></th>
#> </tr>
#> <tr>
#> <th style="empty-cells: hide;border-bottom:hidden;" colspan="1"></th>
#> <th style="border-bottom:hidden;padding-bottom:0; padding-left:3px;padding-right:3px;text-align: center; " colspan="3"><div style="border-bottom: 1px solid #ddd; padding-bottom: 5px; ">Not Hispanic or Latino</div></th>
#> <th style="border-bottom:hidden;padding-bottom:0; padding-left:3px;padding-right:3px;text-align: center; " colspan="3"><div style="border-bottom: 1px solid #ddd; padding-bottom: 5px; ">Hispanic or Latino</div></th>
#> <th style="border-bottom:hidden;padding-bottom:0; padding-left:3px;padding-right:3px;text-align: center; " colspan="3"><div style="border-bottom: 1px solid #ddd; padding-bottom: 5px; ">Unknown/Not Reported Ethnicity</div></th>
#> </tr>
#>   <tr>
#>    <th style="text-align:left;"> Racial
#> Categories </th>
#>    <th style="text-align:right;"> Female </th>
#>    <th style="text-align:right;"> Male </th>
#>    <th style="text-align:right;"> Unknown/<br>Not Reported </th>
#>    <th style="text-align:right;"> Female </th>
#>    <th style="text-align:right;"> Male </th>
#>    <th style="text-align:right;"> Unknown/<br>Not Reported </th>
#>    <th style="text-align:right;"> Female </th>
#>    <th style="text-align:right;"> Male </th>
#>    <th style="text-align:right;"> Unknown/<br>Not Reported </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> American Indian/Alaska Native </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> Asian </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> Native Hawaiian or Other Pacific Islander </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> Black or African American </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 2 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> White </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 2 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 2 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> More than One Race </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> Unknown or Not Reported </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#> </tbody>
#> </table>

table_nih_enrollment(ds_1) %>%
  tidyr::spread(key=gender, value=n)
#> # A tibble: 21 × 5
#>    race                                      ethnicity      Female  Male Unkno…¹
#>    <fct>                                     <fct>           <int> <int>   <int>
#>  1 American Indian/Alaska Native             Not Hispanic …      0     0       0
#>  2 American Indian/Alaska Native             Hispanic or L…      0     0       0
#>  3 American Indian/Alaska Native             Unknown/Not R…      0     0       0
#>  4 Asian                                     Not Hispanic …      0     0       0
#>  5 Asian                                     Hispanic or L…      0     0       0
#>  6 Asian                                     Unknown/Not R…      0     0       0
#>  7 Native Hawaiian or Other Pacific Islander Not Hispanic …      0     0       0
#>  8 Native Hawaiian or Other Pacific Islander Hispanic or L…      0     0       0
#>  9 Native Hawaiian or Other Pacific Islander Unknown/Not R…      0     0       0
#> 10 Black or African American                 Not Hispanic …      0     2       0
#> # … with 11 more rows, and abbreviated variable name ¹​`Unknown/Not Reported`
#> # ℹ Use `print(n = ...)` to see more rows

table_nih_enrollment(ds_1) %>%
  dplyr::mutate(
    gender_ethnicity = paste0(gender, " by ", ethnicity)
  ) %>%
  dplyr::select(-gender, -ethnicity) %>%
  tidyr::spread(key=gender_ethnicity, value=n)
#> # A tibble: 7 × 10
#>   race   Femal…¹ Femal…² Femal…³ Male …⁴ Male …⁵ Male …⁶ Unkno…⁷ Unkno…⁸ Unkno…⁹
#>   <fct>    <int>   <int>   <int>   <int>   <int>   <int>   <int>   <int>   <int>
#> 1 Ameri…       0       0       0       0       0       0       0       0       0
#> 2 Asian        0       0       0       0       0       0       0       0       0
#> 3 Nativ…       0       0       0       0       0       0       0       0       0
#> 4 Black…       0       0       1       0       2       0       0       0       0
#> 5 White        0       1       0       2       2       0       0       0       0
#> 6 More …       0       0       0       0       0       0       0       0       0
#> 7 Unkno…       0       0       0       0       0       0       0       0       0
#> # … with abbreviated variable names ¹​`Female by Hispanic or Latino`,
#> #   ²​`Female by Not Hispanic or Latino`,
#> #   ³​`Female by Unknown/Not Reported Ethnicity`, ⁴​`Male by Hispanic or Latino`,
#> #   ⁵​`Male by Not Hispanic or Latino`,
#> #   ⁶​`Male by Unknown/Not Reported Ethnicity`,
#> #   ⁷​`Unknown/Not Reported by Hispanic or Latino`,
#> #   ⁸​`Unknown/Not Reported by Not Hispanic or Latino`, …

ds_2 <- tibble::tribble(
  ~subject_id,  ~gender , ~race                      , ~ethnicity    ,
           1L,  "Male"  , "Black or African American", "Not Latino"  ,
           2L,  "Male"  , "Black or African American", "Not Latino"  ,
           3L,  "Female", "Black or African American", "Unknown"     ,
           4L,  "Male"  , "White"                    , "Not Latino"  ,
           5L,  "Male"  , "White"                    , "Not Latino"  ,
           6L,  "Female", "White"                    , "Not Latino"  ,
           7L,  "Male"  , "White"                    , "Latino"      ,
           8L,  "Male"  , "White"                    , "Latino"
)

ds_lu_ethnicity <- tibble::tribble(
  ~input      ,   ~displayed                      ,
  "Not Latino",  "Not Hispanic or Latino"         ,
  "Latino"    ,  "Hispanic or Latino"             ,
  "Unknown"   ,  "Unknown/Not Reported Ethnicity"
)
table_nih_enrollment(ds_2, d_lu_ethnicity=ds_lu_ethnicity)
table_nih_enrollment_pretty(ds_2, d_lu_ethnicity=ds_lu_ethnicity)
#> <table class="table table-striped table-hover table-condensed table-responsive" style="width: auto !important; margin-left: auto; margin-right: auto;">
#>  <thead>
#> <tr>
#> <th style="empty-cells: hide;border-bottom:hidden;" colspan="1"></th>
#> <th style="border-bottom:hidden;padding-bottom:0; padding-left:3px;padding-right:3px;text-align: center; " colspan="9"><div style="border-bottom: 1px solid #ddd; padding-bottom: 5px; ">Ethnic Categories</div></th>
#> </tr>
#> <tr>
#> <th style="empty-cells: hide;border-bottom:hidden;" colspan="1"></th>
#> <th style="border-bottom:hidden;padding-bottom:0; padding-left:3px;padding-right:3px;text-align: center; " colspan="3"><div style="border-bottom: 1px solid #ddd; padding-bottom: 5px; ">Not Hispanic or Latino</div></th>
#> <th style="border-bottom:hidden;padding-bottom:0; padding-left:3px;padding-right:3px;text-align: center; " colspan="3"><div style="border-bottom: 1px solid #ddd; padding-bottom: 5px; ">Hispanic or Latino</div></th>
#> <th style="border-bottom:hidden;padding-bottom:0; padding-left:3px;padding-right:3px;text-align: center; " colspan="3"><div style="border-bottom: 1px solid #ddd; padding-bottom: 5px; ">Unknown/Not Reported Ethnicity</div></th>
#> </tr>
#>   <tr>
#>    <th style="text-align:left;"> Racial
#> Categories </th>
#>    <th style="text-align:right;"> Female </th>
#>    <th style="text-align:right;"> Male </th>
#>    <th style="text-align:right;"> Unknown/<br>Not Reported </th>
#>    <th style="text-align:right;"> Female </th>
#>    <th style="text-align:right;"> Male </th>
#>    <th style="text-align:right;"> Unknown/<br>Not Reported </th>
#>    <th style="text-align:right;"> Female </th>
#>    <th style="text-align:right;"> Male </th>
#>    <th style="text-align:right;"> Unknown/<br>Not Reported </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> American Indian/Alaska Native </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> Asian </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> Native Hawaiian or Other Pacific Islander </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> Black or African American </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 2 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> White </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 2 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 2 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> More than One Race </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> Unknown or Not Reported </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#> </tbody>
#> </table>

## Read a 500-patient fake dataset
path <- system.file("misc/example-data-1.csv", package="codified")
ds_3 <- readr::read_csv(path) %>%
  dplyr::mutate(
    gender     = as.character(gender),
    race       = as.character(race),
    ethnicity  = as.character(ethnicity)
  )
#> Rows: 500 Columns: 12
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr  (4): name_last, name_first, address, email
#> dbl  (7): record_id, phone, ethnicity, race, gender, height_cm, weight_kg
#> date (1): dob
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

ds_lu_gender <- tibble::tribble(
  ~input,   ~displayed                      ,
  "0"   ,  "Female",
  "1"   ,  "Male",
  "U"   ,  "Unknown/Not Reported"
)
ds_lu_race <- tibble::tribble(
  ~input ,   ~displayed                      ,
  "1"    , "American Indian/Alaska Native",
  "2"    , "Asian",
  "3"    , "Native Hawaiian or Other Pacific Islander",
  "4"    , "Black or African American",
  "5"    , "White",
  "M"    , "More than One Race",
  "6"    , "Unknown or Not Reported"
)
ds_lu_ethnicity <- tibble::tribble(
  ~input,   ~displayed                      ,
  "2"   ,  "Not Hispanic or Latino"         ,
  "1"   ,  "Hispanic or Latino"             ,
  "0"   ,  "Unknown/Not Reported Ethnicity"
)

table_nih_enrollment(
  d              = ds_3,
  d_lu_gender    = ds_lu_gender,
  d_lu_race      = ds_lu_race,
  d_lu_ethnicity = ds_lu_ethnicity
)

table_nih_enrollment_pretty(
  d              = ds_3,
  d_lu_gender    = ds_lu_gender,
  d_lu_race      = ds_lu_race,
  d_lu_ethnicity = ds_lu_ethnicity
)
#> <table class="table table-striped table-hover table-condensed table-responsive" style="width: auto !important; margin-left: auto; margin-right: auto;">
#>  <thead>
#> <tr>
#> <th style="empty-cells: hide;border-bottom:hidden;" colspan="1"></th>
#> <th style="border-bottom:hidden;padding-bottom:0; padding-left:3px;padding-right:3px;text-align: center; " colspan="9"><div style="border-bottom: 1px solid #ddd; padding-bottom: 5px; ">Ethnic Categories</div></th>
#> </tr>
#> <tr>
#> <th style="empty-cells: hide;border-bottom:hidden;" colspan="1"></th>
#> <th style="border-bottom:hidden;padding-bottom:0; padding-left:3px;padding-right:3px;text-align: center; " colspan="3"><div style="border-bottom: 1px solid #ddd; padding-bottom: 5px; ">Not Hispanic or Latino</div></th>
#> <th style="border-bottom:hidden;padding-bottom:0; padding-left:3px;padding-right:3px;text-align: center; " colspan="3"><div style="border-bottom: 1px solid #ddd; padding-bottom: 5px; ">Hispanic or Latino</div></th>
#> <th style="border-bottom:hidden;padding-bottom:0; padding-left:3px;padding-right:3px;text-align: center; " colspan="3"><div style="border-bottom: 1px solid #ddd; padding-bottom: 5px; ">Unknown/Not Reported Ethnicity</div></th>
#> </tr>
#>   <tr>
#>    <th style="text-align:left;"> Racial
#> Categories </th>
#>    <th style="text-align:right;"> Female </th>
#>    <th style="text-align:right;"> Male </th>
#>    <th style="text-align:right;"> Unknown/<br>Not Reported </th>
#>    <th style="text-align:right;"> Female </th>
#>    <th style="text-align:right;"> Male </th>
#>    <th style="text-align:right;"> Unknown/<br>Not Reported </th>
#>    <th style="text-align:right;"> Female </th>
#>    <th style="text-align:right;"> Male </th>
#>    <th style="text-align:right;"> Unknown/<br>Not Reported </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> American Indian/Alaska Native </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 8 </td>
#>    <td style="text-align:right;"> 10 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> Asian </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> Native Hawaiian or Other Pacific Islander </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 28 </td>
#>    <td style="text-align:right;"> 22 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 3 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> Black or African American </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 5 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 172 </td>
#>    <td style="text-align:right;"> 154 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 14 </td>
#>    <td style="text-align:right;"> 6 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> White </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 22 </td>
#>    <td style="text-align:right;"> 33 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 2 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> More than One Race </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;border-right:1px solid;"> Unknown or Not Reported </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 7 </td>
#>    <td style="text-align:right;"> 7 </td>
#>    <td style="text-align:right;border-right:1px solid;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>   </tr>
#> </tbody>
#> </table>